Bonus

+++ client/src/components/ListComponent.js
@@ -3,6 +3,7 @@ import axios from 'axios'
 
 const ListComponent = () =>{
     const [listState, setListState] = useState([])
+    
     useEffect(()=>{
         axios.get("http://localhost:8000/api/players")
             .then(res => {
@@ -10,12 +11,16 @@ const ListComponent = () =>{
             })
             .catch(err => console.log(err))
     },[])
+
     const deleteHandler = id =>{
-        axios.delete("http://localhost:8000/api/players/"+id)
+        var check = window.confirm("Are you sure?")
+        if(check){
+            axios.delete("http://localhost:8000/api/players/"+id)
             .then(res =>{
                 removeFromDom(id)
             })
             .catch(err => console.log("ERR: ",err))
+        }
     }
     const removeFromDom = id =>{
         setListState(listState.filter(player => player._id !== id))
